% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{censor}
\alias{censor}
\title{Censor}
\usage{
censor(top_table, censor_data = censor_genes, choice = FALSE, greater_than)
}
\description{
Censor is used to remove genes from similarity table that are garbage (too many associations)
}
