% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean colnames}
\usage{
clean_colnames(dataset, summary_df = universal_gene_summary)
}
\arguments{
\item{dataset}{A dataset to fix gene names}

\item{summary_df}{The gene_summary dataframe with all gene symbols}
}
\description{
Clean colnames
}
