% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colorRampPalette_shuffle}
\alias{colorRampPalette_shuffle}
\title{Function to generate shuffled colors based on `colorRampPalette`}
\usage{
colorRampPalette_shuffle(colors, seed = NULL, ...)
}
\arguments{
\item{colors}{Colors to interpolate; must be a valid argument to col2rgb().}
}
\value{
A function with argument a vector of values between 0 and 1 that are
        mapped to a numeric matrix of RGB color values with one row per color
        and 3 or 4 columns
}
\description{
Function to generate shuffled colors based on `colorRampPalette`
}
\examples{
colorRampPalette_shuffle(colors = c("#DEBBF4", "#BE34EF", "#AC7B84"))
colorRampPalette_shuffle(colors = c("#DEBBF4", "#BE34EF", "#AC7B84"), seed = 123L)
}
