% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ddh_colors}
\alias{ddh_colors}
\title{Function to extract DDH colors as hex codes}
\usage{
ddh_colors(...)
}
\arguments{
\item{...}{Character names of colors}
}
\description{
Function to extract DDH colors as hex codes
}
\examples{
ddh_colors()
ddh_colors("gene")

}
