% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ddh_pal_c}
\alias{ddh_pal_c}
\title{Return function to interpolate a continuous DDH color palette}
\usage{
ddh_pal_c(palette = "gene", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ddh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to [grDevices::colorRampPalette()]}
}
\description{
Return function to interpolate a continuous DDH color palette
}
\examples{
ddh_pal_c()(10)
ddh_pal_c("protein")(3)
ddh_pal_c("cOMPOUND", reverse = TRUE)(5)

}
