% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ddh_pal_d}
\alias{ddh_pal_d}
\title{Return function to interpolate a discrete DDH color palette}
\usage{
ddh_pal_d(palette = "gene", reverse = FALSE, shuffle = FALSE, seed = NULL)
}
\arguments{
\item{palette}{Character name of palette in ddh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
Return function to interpolate a discrete DDH color palette
}
\examples{
ddh_pal_d()(10)
ddh_pal_d("protein")(3)
ddh_pal_d("cOMPOUND", reverse = TRUE)(5)

}
