% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{download_ddh_data}
\alias{download_ddh_data}
\title{Function to download all DDH .Rds files}
\usage{
download_ddh_data(
  app_data_dir,
  object_name = NULL,
  test = FALSE,
  overwrite = FALSE,
  privateMode = TRUE,
  raw = FALSE
)
}
\arguments{
\item{app_data_dir}{Data directory path.}

\item{object_name}{Optional object name to get a single file; default gets all files in bucket}

\item{test}{Boolean that will download test data instead of full data}

\item{overwrite}{Boolean that deletes the app_data_dir thereby removing files, before remaking dir and downloading all files}

\item{privateMode}{Boolean indicating if private data is required.}
}
\description{
Function to download all DDH .Rds files
}
