% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fix_names}
\alias{fix_names}
\title{Fix names}
\usage{
fix_names(wrong_name, summary_df = universal_gene_summary)
}
\arguments{
\item{wrong_name}{A gene name that is not its official symbol}

\item{summary_df}{The gene_summary dataframe with all gene symbols}
}
\description{
Fix names
}
