% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_location}
\alias{gene_location}
\title{gene_location}
\format{
A data frame with 20388 rows and 14 variables:
\describe{
  \item{ensembl_transcript_id}{Ensembl transcript id}
  \item{ensembl_gene_id}{Ensembl gene id}
  \item{chromosome_name}{Chromosome name}
  \item{transcript_start}{Transcript start site}
  \item{transcript_end}{Transcript end site}
  \item{strand}{Strand}
  \item{band}{Band}
  \item{approved_symbol}{Approved gene symbol}
  \item{percentage_gene_gc_content}{Percent GC content}
  \item{gene_name}{Gene name}
  \item{description}{Gene description}
  \item{cds_length}{CDS length}
}
}
\source{
\url{http://uswest.ensembl.org/index.html}
}
\usage{
data(gene_location)
}
\description{
A dataset containing information of human genes and their location across the human genome
}
\keyword{datasets}
