% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gene_molecular_features_barplot}
\alias{gene_molecular_features_barplot}
\title{Molecular Features Bar Plot}
\usage{
gene_molecular_features_barplot(input = list(), n_features = 20, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a barplot. If an error is thrown, then will return a bomb plot.
}
\description{
Molecular Features Bar Plot
}
\examples{
gene_molecular_features_barplot(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
gene_molecular_features_barplot(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
gene_molecular_features_barplot(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')), n_features = 10)
}
