% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gene_molecular_features_pathway_barplot}
\alias{gene_molecular_features_pathway_barplot}
\title{Molecular Features Pathways Bar Plot}
\usage{
gene_molecular_features_pathway_barplot(input = list(), n_pathways = 10, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a barplot. If an error is thrown, then will return a bomb plot.
}
\description{
Molecular Features Pathways Bar Plot
}
\examples{
gene_molecular_features_pathway_barplot(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
gene_molecular_features_pathway_barplot(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
gene_molecular_features_pathway_barplot(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')), n_pathways = 20)
}
