% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_cluster}
\alias{get_cluster}
\title{Clustering Table}
\usage{
get_cluster(input = list())
}
\arguments{
\item{input}{Expecting a list containing a content variable.}
}
\value{
Returns a vector containing the cluster number.
}
\description{
This is a helper function that takes a gene name and return a vector containing the cluster numbers for input
}
\examples{
get_cluster(input = list(type = 'gene', content = 'ROCK1'))
get_cluster(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
\dontrun{
get_cluster(input = list(type = 'gene', content = 'ROCK1'))
}
}
