% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_content}
\alias{get_content}
\title{Function to load data from AWS into environment}
\usage{
get_content(object_name, dataset = FALSE)
}
\arguments{
\item{dataset}{Boolean that will indicate if it is a dataset and therefore fetch from _data/}

\item{object_names}{String containing file name to get}
}
\description{
Function to load data from AWS into environment
}
\examples{
get_content("ROCK1")
get_content(object_name = "ROCK2")
get_content(object_name = "gene_band_boundaries", dataset = TRUE)
\dontrun{
get_content("ROCK1")
}
}
