% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_data_object}
\alias{get_data_object}
\title{Function to get filtered data object from environment}
\usage{
get_data_object(object_names, dataset_name = NULL, pivotwider = FALSE)
}
\arguments{
\item{object_names}{Character vector, can be greater than 1, of file names to get}

\item{data_set_name}{String containing dataset name to filter out from object}
}
\description{
Function to get filtered data object from environment
}
\examples{
ROCK1 <- get_data_object(object_names = c("ROCK1"))
get_data_object(object_names = c("ROCK1"), dataset_name = "gene_female_tissue")
get_data_object(object_names = c("ROCK1", "ROCK2"), dataset_name = "gene_female_tissue")
get_data_object(object_names = c("ROCK1", "ROCK2"), dataset_name = "gene_female_tissue", pivotwider = TRUE)
get_data_object(object_names = c("ROCK1", "ROCK2"), dataset_name = "gene_subcell")
get_data_object(object_names = c("ROCK2"), dataset_name = "gene_subcell", pivotwider = TRUE)
get_data_object(object_names = c("ROCK1", "ROCK2"), dataset_name = "gene_subcell", pivotwider = TRUE)
\dontrun{
get_data_object(object_name = c("ROCK1"), dataset_name = "gene_female_tissue")
}
}
