% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{get_essential}
\alias{get_essential}
\title{Get Essential}
\usage{
get_essential(input = list())
}
\description{
Gets number of cells lines a gene is essential within
}
\examples{
get_essential(input = list(type = "gene", content = "ROCK1"))
get_essential(input = list(type = "gene", content = "ROCK2"))
get_essential(input = list(type = "gene", content = c("ROCK2", "ROCK1")))
get_essential(input = list(type = "gene", subtype = "custom_gene_list", content = c("RDX", "ROCK2", "DTX3L", "MSN", "SORL1", "EZR")))
get_essential(input = list(type = "cell", content = "HEPG2"))
get_essential(input = list(type = "cell", content = c("HEPG2", "HUH7")))
get_essential(input = list(type = "compound", content = "aspirin"))
}
