% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_gene_symbols_for_pathway}
\alias{get_gene_symbols_for_pathway}
\title{Pathway Genes}
\usage{
get_gene_symbols_for_pathway(pathway_id)
}
\description{
\code{make_pathway_genes} returns a vector of gene symbols in a queried pathway
}
