% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_stats}
\alias{get_stats}
\title{Get Stats}
\usage{
get_stats(
  data_universal_stats_summary = universal_stats_summary,
  data_set,
  var
)
}
\arguments{
\item{data_universal_stats_summary}{Dataframe of stats summary generated in ddh_data.}

\item{data_set}{A character indicating data set from which the stats were generated, typically one of achilles, expression_gene, expression_protein, or prism name.}

\item{var}{A character indicating the variable to extract from the stats summary dataframe, typically one of threshold, sd, mean, upper, pr lower.}
}
\description{
Get Stats
}
\examples{
get_stats(data_set = "achilles", var = "sd")
}
