% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_temporary_url}
\alias{get_temporary_url}
\title{Create a temporary URL to download a file from a S3 bucket}
\usage{
get_temporary_url(bucket_name_env, key)
}
\arguments{
\item{bucket_name_env}{A character the name of an environment variable containing the bucket name}

\item{key}{A character the file within the bucket}
}
\description{
Create a temporary URL to download a file from a S3 bucket
}
