% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{good_file_namer}
\alias{good_file_namer}
\title{Good File Namer}
\usage{
good_file_namer(input = list)
}
\arguments{
\item{input}{Expecting a list that contains a content object}
}
\description{
Good File Namer
}
\examples{
good_file_namer(input = list(content = c("ROCK1", "ROCK2")))
}
