% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{help_extract}
\alias{help_extract}
\title{Extract Documentation Sections}
\usage{
help_extract(fun, section = "Description", ...)
}
\arguments{
\item{fun}{Function name.}

\item{section}{Help section to extract.}
}
\description{
Extract parts of a function help documentation.
}
\examples{
help_extract(make_radial, package = ddh, section = "Description")
help_extract(make_radial, package = ddh, section = "Examples")
}
