% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_ddh_db}
\alias{load_ddh_db}
\title{Function to load all DDH db connections}
\usage{
load_ddh_db(object_name = NULL)
}
\arguments{
\item{object_name}{Optional object name to load a single db connection; default loads all connections}
}
\description{
Function to load all DDH db connections
}
