% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_pdb}
\alias{load_pdb}
\title{Load single PDB}
\usage{
load_pdb(app_data_dir, input = list())
}
\description{
Load single PDB
}
\examples{
load_pdb(input = list(content = c("ROCK1")))
load_pdb(input = list(content = c("ROCK3")))
load_pdb(input = list(content = c("ROCK1", "ROCK2")))
}
