% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_bipartite_graph}
\alias{make_bipartite_graph}
\title{Bipartite Graph Graph}
\usage{
make_bipartite_graph(
  input = list(),
  censor = character(),
  collapsed = TRUE,
  threshold = 10,
  corr_type = "positive",
  card = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a bipartite graph graph. If an error is thrown, then will return an empty graph.
}
\description{
\code{make_bipartite_graph} returns an image of ...
}
\details{
This is a graph function that takes a gene name and returns a bipartite graph graph
}
\examples{
make_bipartite_graph(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_bipartite_graph(input = list(type = "gene", content = "ROCK1"), collapsed = TRUE, threshold = 10, corr_type = "positive")
make_bipartite_graph(input = list(type = "gene", content = c("ROCK1", "ROCK2")), collapsed = TRUE, threshold = 10, corr_type = "positive", censor = c("ADP", "Adenosine triphosphate"))
make_bipartite_graph(input = list(type = "compound", content = "adp"))
\dontrun{
make_bipartite_graph(input = list(type = 'gene', content = 'ROCK1'))
}
}
