% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_bomb_plot}
\alias{make_bomb_plot}
\title{Bomb Plot}
\usage{
make_bomb_plot()
}
\value{
If no error, then returns a bomb plot plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_bomb_plot} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a bomb plot plot
}
\examples{
make_bomb_plot()
\dontrun{
make_bomb_plot()
}
}
