% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_bottom_table}
\alias{make_bottom_table}
\title{Bottom Table}
\usage{
make_bottom_table(input = list(), gls = FALSE, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a bottom Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_bottom_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a bottom Table
}
\examples{
make_bottom_table(input = list(type = 'gene', content = 'ROCK1'))
make_bottom_table(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
\dontrun{
make_bottom_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
