% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cca_genes}
\alias{make_cca_genes}
\title{Gene Pathways Bar Plot}
\usage{
make_cca_genes(input = list(), n_features = 10, gset = NULL, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a barplot. If an error is thrown, then will return a bomb plot.
}
\description{
Gene Pathways Bar Plot
}
\examples{
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), gset = "GOBP")
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), gset = c("GOBP", "C2"))
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')), n_features = 20)
}
