% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cca_genes_table}
\alias{make_cca_genes_table}
\title{Gene-Pathway CCA Table}
\usage{
make_cca_genes_table(input = list(), gene_set = NULL, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a gene-pathway table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a gene-pathway table
}
\examples{
make_cca_genes_table(input = list(type = 'gene', content = 'ROCK1'))
make_cca_genes_table(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
\dontrun{
make_cca_genes_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
