% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cca_pathway_table}
\alias{make_cca_pathway_table}
\title{Pathway-Pathway CCA Table}
\usage{
make_cca_pathway_table(input = list(), gene_set = NULL, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a pathway-pathway table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a pathway name and returns a pathway-pathway table
}
\examples{
make_cca_pathway_table(input = list(type = 'pathway', content = 'GOBP_FIBROBLAST_GROWTH_FACTOR_PRODUCTION'))
\dontrun{
make_cca_pathway_table(input = list(type = 'pathway', content = 'GOBP_FIBROBLAST_GROWTH_FACTOR_PRODUCTION'))
}
}
