% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cca_pathways}
\alias{make_cca_pathways}
\title{Co-essentiality Pathway Plot}
\usage{
make_cca_pathways(
  input = list(),
  n_features = 10,
  gset = NULL,
  facet_by_geneset = FALSE,
  ...
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a barplot. If an error is thrown, then will return a bomb plot.
}
\description{
Top pathway associations for the queried pathway/s. The x-axis shows the first canonical correlation between pathways (range 0-1).
}
\examples{
make_cca_pathways(input = list(type = 'pathway', query = 5887))
make_cca_pathways(input = list(type = 'pathway', query = 5887), gset = "GOBP")
make_cca_pathways(input = list(type = 'pathway', query = 5887), gset = c("GOBP", "C2"))
}
