% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cell_image}
\alias{make_cell_image}
\title{Cell Images}
\usage{
make_cell_image(input = list(), card = FALSE)
}
\description{
Cells image shown at low (top) and high (bottom) growth density.
}
\examples{
make_cell_image(input = list(content = "HEPG2"))
}
