% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cell_line_table}
\alias{make_cell_line_table}
\title{Cell Summary Table}
\usage{
make_cell_line_table(
  data_cell_expression_meta = cell_expression_meta,
  input = list()
)
}
\description{
Cell Summary Table
}
\examples{
make_cell_line_table(input = list(content = 'HEPG2'))
}
