% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cell_similarity}
\alias{make_cell_similarity}
\title{Cell Similarity Plot}
\usage{
make_cell_similarity(
<<<<<<< HEAD
  data_cell_dependency_sim = cell_dependency_sim,
  data_cell_expression_sim = cell_expression_sim,
=======
  data_cell_line_dep_sim = cell_line_dep_sim,
  data_cell_line_exp_sim = cell_line_exp_sim,
>>>>>>> main
  similarity = "dependency",
  input = list(),
  card = FALSE,
  scale = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cell similarity plot If an error is thrown, then will return a bomb plot.
}
\description{
Each point shows the ranked linear model coefficient estimate value ordered from high to low for each query.
}
\examples{
make_cell_similarity(input = list(type = "cell", query = "HEPG2", content = "HEPG2"))
make_cell_similarity(input = list(type = "cell", query = "HEPG2", content = "HEPG2"), similarity = "expression")
\dontrun{
make_cell_similarity(input = list(type = "cell", query = "HEPG2", content = "HEPG2"))
}
}
