% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cellanatogram_table}
\alias{make_cellanatogram_table}
\title{Cellanatogram Table}
\usage{
make_cellanatogram_table(data_gene_subcell = gene_subcell, input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a cellanatogram Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_cellanatogram_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a cellanatogram Table
}
\examples{
make_cellanatogram_table(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_cellanatogram_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
