% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellanatogramfacet}
\alias{make_cellanatogramfacet}
\title{Cellanatogramfacet Plot}
\usage{
make_cellanatogramfacet(cellanatogram_data = subcell, input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cellanatogramfacet plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_cellanatogramfacet} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a cellanatogramfacet plot
}
\examples{
make_cellanatogramfacet(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cellanatogramfacet(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
make_cellanatogramfacet(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
\dontrun{
make_cellanatogramfacet(input = list(type = 'gene', content = 'ROCK1'))
}
}
