% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellbar}
\alias{make_cellbar}
\title{Cell Dependencies Bar Plot}
\usage{
make_cellbar(
  celldeps_data = achilles_long,
  prism_data = prism_long,
  expression_data = expression_names,
  mean_achilles = mean_virtual_achilles,
  mean_prism = mean_virtual_prism_cor,
  mean_cell_line = mean_virtual_achilles_cell_line,
  input = list(),
  card = FALSE,
  scale = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cellbar plot. If an error is thrown, then will return a bomb plot.
}
\description{
Each bar shows the dependency scores of the queried genes in a cell line. Dependency scores less than -1 indicate a gene that is essential within a cell line. Dependency scores close to 0 mean no changes in fitness when the gene is knocked out. Dependency scores greater than 1 indicate gene knockouts lead to a gain in fitness.
}
\examples{
make_cellbar(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cellbar(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cellbar(input = list(type = 'gene', content = 'ROCK1'))
}
}
