% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellbins}
\alias{make_cellbins}
\title{Cell Dependencies Density Plot}
\usage{
make_cellbins(input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cellbins plot. If an error is thrown, then will return a bomb plot.
}
\description{
Kernel density estimate of dependency scores. Dependency scores across all cell lines for queried genes, revealing overall influence of a gene on cellular fitness. The interval indicates the 95 percent quantile of the data, the dot indicates the median dependency score. The gray background highlights weak dependency values between -1 and 1.
}
\examples{

make_cellbins(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cellbins(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
make_cellbins(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cellbins(input = list(type = 'gene', content = 'ROCK1'))
}
}
