% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_celldeps}
\alias{make_celldeps}
\title{Dependency Curve Plot}
\usage{
make_celldeps(input = list(), card = FALSE, lineplot = FALSE, scale = NULL)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a celldeps plot. If an error is thrown, then will return a bomb plot.
}
\description{
Each point shows the ranked dependency score ordered from low to high scores. Dependency scores less than -1 indicate a gene that is essential within a cell line. Dependency scores close to 0 mean no changes in fitness when the gene is knocked out. Dependency scores greater than 1 indicate gene knockouts lead to a gain in fitness.
}
\examples{
make_celldeps(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_celldeps(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
make_celldeps(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_celldeps(input = list(type = 'gene', content = 'ROCK1'))
}
}
