% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellgeneprotein}
\alias{make_cellgeneprotein}
\title{Gene versus protein in a cell}
\usage{
make_cellgeneprotein(
  expression_data = expression_long,
  expression_join = expression_names,
  input = list(),
  card = FALSE
)
}
\description{
@importFrom magrittr %>%
}
