% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_censor_table}
\alias{make_censor_table}
\title{Censor}
\usage{
make_censor_table(input = list(), censor = FALSE, greater_than = 300)
}
\description{
Censor is used to remove genes from similarity table that are garbage (too many associations). It does this by filtering out genes above (greater_than) a provided threshold.
}
\examples{
make_censor_table(input = list(type = 'gene', content = 'ROCK1'))
make_censor_table(input = list(type = 'gene', content = 'ROCK1'), choice = TRUE)
make_censor_table(input = list(type = 'gene', content = 'ROCK1'), choice = TRUE, greater_than = 100)
\dontrun{
make_censor_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
