% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cluster_enrich}
\alias{make_cluster_enrich}
\title{Cluster Enrichment Plot}
\usage{
make_cluster_enrich(
  data_gene_signature_clusters = gene_signature_clusters,
  data_gene_signature_cluster_enrichment = gene_signature_cluster_enrichment,
  input = list(),
  ontology = "BP",
  num_terms = 20
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cluster enrich plot. If an error is thrown, then will return a bomb plot.
}
\description{
Enriched GO terms (BP, MF, and CC) by all genes in the selected amino acid signature cluster. The x-axis shows the number of genes in the cluster that belong to each term while the color scale represents the p-values of each enriched term.
}
\examples{
make_cluster_enrich(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cluster_enrich(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cluster_enrich(input = list(type = 'gene', content = 'ROCK1'))
}
}
