% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cluster_enrichment_table}
\alias{make_cluster_enrichment_table}
\title{Clustering Enrichment Table}
\usage{
make_cluster_enrichment_table(input = list(), ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a clustering enrichment Table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a clustering enrichment Table
}
\examples{
make_cluster_enrichment_table(input = list(type = 'gene', content = 'ROCK1'))
make_cluster_enrichment_table(input = list(type = 'gene', content = 'ROCK1'), ontology = "MF")
make_cluster_enrichment_table(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
\dontrun{
make_cluster_enrichment_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
