% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_clustering_enrichment_table}
\alias{make_clustering_enrichment_table}
\title{Clustering Enrichment Table}
\usage{
make_clustering_enrichment_table(
  data_gene_signature_clusters = gene_signature_clusters,
  data_gene_signature_cluster_enrichment = gene_signature_cluster_enrichment,
  input = list(),
  ontology = "BP",
  filter_noise = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a clustering enrichment Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_clustering_enrichment_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a clustering enrichment Table
}
\examples{
make_clustering_enrichment_table(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_clustering_enrichment_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
