% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_correlation}
\alias{make_correlation}
\title{Co-essentiality Correlation Plot}
\usage{
make_correlation(
  table_data = achilles_cor_nest,
  cell_line_data = achilles_cell_line_cor_nest,
  prism_data = prism_cor_nest,
  input = list(),
  card = FALSE,
  scale = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a correlation plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_correlation} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a correlation plot
}
\examples{
make_correlation(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_correlation(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_correlation(input = list(type = 'gene', content = 'ROCK1'))
}
}
