% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_dep_table}
\alias{make_dep_table}
\title{Dependency Table}
\usage{
make_dep_table(input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a dep Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_dep_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a dep Table
}
\examples{
make_dep_table(input = list(type = 'gene', content = 'ROCK1'))
make_dep_table(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
make_dep_table(input = list(type = 'compound', content = 'aspirin'))
make_dep_table(input = list(type = 'cell', content = 'HEPG2'))
\dontrun{
make_dep_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
