% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_drug_genes_cor_table}
\alias{make_drug_genes_cor_table}
\title{Drug Genes Cor Table}
\usage{
make_drug_genes_cor_table(table_data = drug_genes_cor_table, drug)
}
\description{
\code{make_drug_genes_cor_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a drug genes cor Table
}
