% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_empty_graph}
\alias{make_empty_graph}
\title{Empty Graph Graph}
\usage{
make_empty_graph(type = "gene")
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a empty graph graph. If an error is thrown, then will return an empty graph.
}
\description{
\code{make_empty_graph} returns an image of ...
}
\details{
This is a graph function that takes a gene name and returns a empty graph graph
}
\examples{
make_empty_graph()
\dontrun{
make_empty_graph()
}
}
