% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_empty_plot}
\alias{make_empty_plot}
\title{Function to create an empty}
\usage{
make_empty_plot()
}
\value{
A ggplot2 object.
}
\description{
Function to create an empty
}
\examples{
make_empty_plot()
}
