% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_empty_table}
\alias{make_empty_table}
\title{Function to create an empty table}
\usage{
make_empty_table()
}
\value{
A data.frame.
}
\description{
Function to create an empty table
}
\examples{
make_empty_table()
}
