% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_enrichment_bottom}
\alias{make_enrichment_bottom}
\title{Enrichment Bottom Table}
\usage{
make_enrichment_bottom(
  data_gene_master_negative = gene_master_negative,
  input = list()
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a enrichment bottom table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_enrichment_bottom} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a enrichment bottom table
}
\examples{
make_enrichment_bottom(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_enrichment_bottom(input = list(type = 'gene', content = 'ROCK1'))
}
}
