% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_enrichment_top}
\alias{make_enrichment_top}
\title{Enrichment Top Table}
\usage{
<<<<<<< HEAD
make_enrichment_top(
  data_gene_master_positive = gene_master_positive,
  input = list()
)
=======
make_enrichment_top(data_master_positive = master_positive, input = list())
>>>>>>> main
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a enrichment top table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_enrichment_top} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a enrichment top table
}
\examples{
make_enrichment_top(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_enrichment_top(input = list(type = 'gene', content = 'ROCK1'))
}
}
