% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_example}
\alias{make_example}
\title{Make Example}
\usage{
make_example(privateMode = TRUE)
}
\arguments{
\item{privateMode}{Boolean indicating if private data is pointed to}
}
\value{
An HTML string.
}
\description{
Function to create an HTML string of examples to use on the index page and the start here methods doc
}
\examples{
make_example()
\dontrun{
make_example()
}
}
