% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_expression_table}
\alias{make_expression_table}
\title{Expression Table}
\usage{
make_expression_table(
  expression_data = expression_long,
  expression_join = expression_names,
  input = list(),
  var = "gene"
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a expression Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_expression_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a expression Table
}
\examples{
make_expression_table(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
\dontrun{
make_expression_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
