% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_functional_cell}
\alias{make_functional_cell}
\title{Differential Pathway Expression Plot}
\usage{
make_functional_cell(
  pathway_data = pathways,
  expression_data = expression_long,
  expression_metadata = expression_meta,
  input = list(),
  num_pathways = 10,
  nwords = 5,
  num_genes = 2,
  remove_equivalent_pathways = FALSE,
  card = FALSE
)
}
\description{
The colored vertical bars indicate the pathway median expression for the queried cell line/s while the background grey points indicate the pathway median expression of all the other cell lines. If the query includes only one cell line, the difference between the median pathway expression of that cell line and the median pathway expression of all the other cell lines will be computed and the pathways with higher differences will appear first in the plot. Otherwise, the biggest difference between pathway medians of the queried cell lines will be used to rank the pathways in the plot. Those pathways with higher differences will appear first in the plot.
}
\examples{
make_functional_cell(input = list(type = "cell", query = "HEPG2", content = "HEPG2"))
make_functional_cell(input = list(type = "cell", query = c("HEPG2", "HEL"), content = c("HEPG2", "HEL")))
make_functional_cell(input = list(type = "cell", query = "HEPG2", content = "HEPG2"), card = TRUE)
\dontrun{
make_functional_cell(input = list(type = "cell", query = "HEPG2", content = "HEPG2"))
}
}
