% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_gene_dependency_enrichment_table}
\alias{make_gene_dependency_enrichment_table}
\title{Enrichment Top Table}
\usage{
make_gene_dependency_enrichment_table(input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a enrichment top table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a enrichment top table
}
\examples{
make_gene_dependency_enrichment_table(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_gene_dependency_enrichment_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
