% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_gene_drugs_cor_table}
\alias{make_gene_drugs_cor_table}
\title{Gene Drugs Cor Table}
\usage{
make_gene_drugs_cor_table(input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a gene drugs cor Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_gene_drugs_cor_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a gene drugs cor Table
}
\examples{
make_gene_drugs_cor_table(input = list(type = 'gene', content = 'ACOT4'))
\dontrun{
make_gene_drugs_cor_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
