% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_gene_drugs_table}
\alias{make_gene_drugs_table}
\title{Gene Drugs Table}
\usage{
make_gene_drugs_table(table_data = gene_drugs_table, input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a gene drugs Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_gene_drugs_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a gene drugs Table
}
\examples{
make_gene_drugs_table(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_gene_drugs_table_raw(input = list(content = c("ROCK1", "RDX")))
make_gene_drugs_table_raw(input = list(content = c("ROCK1", "RDX", "ROCK")))
\dontrun{
make_gene_drugs_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
