% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_gene_molecular_features_pathways}
\alias{make_gene_molecular_features_pathways}
\title{Molecular Features Pathways Table}
\usage{
make_gene_molecular_features_pathways(input = list(), ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a molecular features pathways table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a molecular features pathways table
}
\examples{
make_gene_molecular_features_pathways(input = list(type = 'gene', content = 'ROCK1'))
make_gene_molecular_features_pathways(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
\dontrun{
make_gene_molecular_features_pathways(input = list(type = 'gene', content = 'ROCK1'))
}
}
