% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_gene_pathways_components}
\alias{make_gene_pathways_components}
\title{Gene-Pathway and Pathway-Pathway Co-essentiality Table}
\usage{
make_gene_pathways_components(
  data_gene_pathways_components = gene_pathways_components,
  input = list(),
  cutoff = 0.7,
  sign = "positive"
)
}
\arguments{
\item{input}{Expecting a list containing content variable.}

\item{cutoff}{Absolute Pearson's correlation value to filter}

\item{sign}{Character vector. Either "positive" or "negative" to return positive co-essentialities or negative co-essentialities, respectively.}
}
\value{
If no error, then returns a table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene/s or a pathway/s and returns its co-essential pathways and genes
}
\examples{
make_gene_pathways_components(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_gene_pathways_components(input = list(type = 'gene', content = 'ROCK1'))
}
}
