% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_gene_pathways_components_network}
\alias{make_gene_pathways_components_network}
\title{Gene-Pathway co-essentiality plot}
\usage{
make_gene_pathways_components_network(
  data_universal_achilles_long = universal_achilles_long,
  input = list(),
  highlight = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing content variable.}
}
\value{
If no error, then returns a network plot. If an error is thrown, then will return a bomb plot.
}
\description{
Gene-Pathway co-essentiality plot
}
\examples{
make_gene_pathways_components_network(input = list(content = 'ROCK1'))
\dontrun{
make_gene_pathways_components_network(input = list(content = 'ROCK1'))
}
}
