% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_ideogram}
\alias{make_ideogram}
\title{Ideogram Plot}
\usage{
make_ideogram(input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns an ideogram plot. If an error is thrown, then will return a bomb plot
}
\description{
Each point shows the location of the query gene(s) on human chromosomes.
}
\examples{
make_ideogram(input = list(type = "gene", query = "ROCK1", content = "ROCK1"))
make_ideogram(input = list(type = "gene", query = "ROCK1", content = "ROCK1"), card = TRUE)
\dontrun{
make_ideogram(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
}
}
