% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_legend}
\alias{make_legend}
\title{Make Legend}
\usage{
make_legend(fun, html = FALSE, ...)
}
\arguments{
\item{fun}{Function name.}

\item{html}{Boolean indicating HTML format or not. Default is FALSE.}
}
\description{
Make Legend
}
\examples{
make_legend(fun = "make_radial")
}
