% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_metadata_cell}
\alias{make_metadata_cell}
\title{Lineage Similarity Plot}
\usage{
make_metadata_cell(
  data_cell_dependency_sim = cell_dependency_sim,
  data_cell_expression_sim = cell_expression_sim,
  input = list(),
  cell_line_similarity = "dependency",
  metadata = "lineage",
  bonferroni_cutoff = 0.05,
  card = FALSE
)
}
\description{
Similar and dissimilar lineages (and sublineages) associated with the queried cell line/s.
}
\examples{
make_metadata_cell(input = list(type = "cell", content = c("HEPG2")))
make_metadata_cell(input = list(type = "cell", content = c("HEPG2")), cell_line_similarity = "expression")
make_metadata_cell(input = list(type = "cell", content = c("HEPG2")), metadata = "sublineage")
make_metadata_cell(input = list(type = "cell", content = c("HEPG2", "HEL")))
make_metadata_cell(input = list(type = "cell", content = c("HEPG2")), card = TRUE)
\dontrun{
make_metadata_cell(input = list(type = "cell", content = c("HEPG2")))
}
}
