% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_molecular_features_boxplots}
\alias{make_molecular_features_boxplots}
\title{Molecular Features Boxplot}
\usage{
make_molecular_features_boxplots(
  input = list(),
  target_genes = NULL,
  sex_select = NULL,
  lineage_select = NULL,
  lineage_subtype_select = NULL,
  card = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a boxplot. If an error is thrown, then will return a bomb plot.
}
\description{
Molecular features associated with the query ablation.
}
\examples{
make_molecular_features_boxplots(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_molecular_features_boxplots(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
}
