% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_molecular_features_segments}
\alias{make_molecular_features_segments}
\title{Molecular Features Segments Plot}
\usage{
make_molecular_features_segments(input = list(), ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a scatterplot. If an error is thrown, then will return a bomb plot.
}
\description{
Molecular Features Segments Plot
}
\examples{
make_molecular_features_segments(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_molecular_features_segments(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
}
