% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_molecule_structure}
\alias{make_molecule_structure}
\title{Molecule Structure Plot}
\usage{
make_molecule_structure(
  input = list(),
  model_width = 800,
  model_height = 800,
  card = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a molecule structure plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_molecule_structure} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a molecule structure plot
}
\examples{
make_molecule_structure(input = list(content = "aspirin"))
make_molecule_structure(input = list(content = "ibuprofen"))
make_molecule_structure(input = list(content = 2244))
\dontrun{
make_molecule_structure(input = list(content = "aspirin"))
}
}
