% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_pathway_table}
\alias{make_pathway_table}
\title{Pathway List Table}
\usage{
make_pathway_table(input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a pathway list table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a sub-table of gene sets and pathways that contain your gene query
}
\examples{
make_pathway_table(input = list(type = 'gene', subtype = 'gene', content = 'ROCK1'))
make_pathway_table(input = list(type = 'gene', subtype = 'gene', content = c('ROCK1', 'ROCK2')))
make_pathway_table(input = list(type = 'gene', subtype = 'pathway', query = 5887))
}
