% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_protein_domain_plot}
\alias{make_protein_domain_plot}
\title{Protein Domain Plot}
\usage{
make_protein_domain_plot(
  input = list(),
  domain_data = protein_domains,
  dom_var = NULL,
  ptm_var = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a protein domain plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_protein_domain_plot} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a protein domain plot plot
}
\examples{
make_protein_domain_plot(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_protein_domain_plot(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_protein_domain_plot(input = list(type = 'gene', content = 'ROCK1'))
}
}
