% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_protein_function_predictions_table}
\alias{make_protein_function_predictions_table}
\title{Predicted Protein Functions}
\usage{
make_protein_function_predictions_table(input = list())
}
\arguments{
\item{input}{Expecting a list containing a content variable.}
}
\value{
Returns a table with the predicted protein functions.
}
\description{
Predicted Protein Functions
}
\examples{
make_protein_function_predictions_table(input = list(content = 'ROCK1'))
make_protein_function_predictions_table(input = list(content = c('ROCK1', 'ROCK2')))
}
