% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_protein_sequence}
\alias{make_protein_sequence}
\title{Make Protein Sequence}
\usage{
make_protein_sequence(input = list(), ...)
}
\arguments{
\item{input}{A list containing a content variable.}

\item{var}{Variable that determines which text is returned}
}
\description{
The make_protein_sequence function takes a gene as an input and returns its protein sequence.
}
\examples{
make_protein_sequence(input = list(type = "gene", content = c("ROCK1")))
make_protein_sequence(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
}
