% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_pubmed_table}
\alias{make_pubmed_table}
\title{Pubmed Table}
\usage{
make_pubmed_table(input = list(), n_papers = NULL, title = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a pubmed table..
}
\description{
\code{make_pubmed_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a pubmed Table
}
\examples{
make_pubmed_table(input = list(type = 'gene', content = 'ROCK1'))
make_pubmed_table(input = list(type = 'gene', content = 'ROCK1'), title = TRUE)
\dontrun{
make_pubmed_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
