% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_radial_bar}
\alias{make_radial_bar}
\title{Amino Acid Bar Plot}
\usage{
make_radial_bar(input = list(), relative = TRUE, cluster = FALSE, card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a bar plot. If an error is thrown, then will return a bomb plot.
}
\description{
Amino acid signature/s (percentage of each amino acid in a protein) of the queried gene/clusters versus the mean amino acid signature of all the other proteins in the dataset (N = 20375).
}
\examples{
make_radial_bar(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_radial_bar(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), cluster = TRUE)
make_radial_bar(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_radial_bar(input = list(type = 'gene', content = 'ROCK1'))
}
}
