% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_roxygen}
\alias{make_roxygen}
\title{Function to create roxygen structures}
\usage{
make_roxygen(fun_name, type = "graph")
}
\arguments{
\item{fun_name}{Function name.}

\item{type}{Function type.}
}
\value{
A roxygen structure.
}
\description{
Function to create roxygen structures
}
\examples{
make_roxygen("make_ideogram")
}
