% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_signature_clusters_table}
\alias{make_signature_clusters_table}
\title{Clustering Table}
\usage{
make_signature_clusters_table(input = list(), ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a clustering Table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a clustering Table
}
\examples{
make_signature_clusters_table(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_signature_clusters_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
