% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_structure3d}
\alias{make_structure3d}
\title{Protein 3D Structure Plot}
\usage{
make_structure3d(
  pdb_ids = uniprot_pdb_table,
  protein_data = proteins,
  app_data_dir = NULL,
  gene_id = NULL,
  pdb_id = NULL,
  input = list(),
  color = FALSE,
  ribbon = FALSE,
  selection = FALSE,
  resi = 1:10,
  chain = "A",
  resn = NULL,
  invert = NULL,
  elem = NULL
)
}
\description{
Protein 3D predicted structure rendered in an interactive ribbon diagram.
}
