% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_structure3d_table}
\alias{make_structure3d_table}
\title{Plot for 3D protein structure table}
\usage{
make_structure3d_table(
<<<<<<< HEAD
  data_gene_uniprot_pdb_table = gene_uniprot_pdb_table,
  data_universal_proteins = universal_proteins,
=======
  data_uniprot_pdb_table = uniprot_pdb_table,
  data_hmdb_proteins = proteins,
>>>>>>> main
  input = list()
)
}
\description{
Plot for 3D protein structure table
}
\examples{
make_structure3d_table(input = list(content = 'ROCK1'))
}
