% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_sublineage}
\alias{make_sublineage}
\title{Dependency Sub-lineage Plot}
\usage{
make_sublineage(input = list(), card = FALSE, highlight = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a lineage plot. If an error is thrown, then will return a bomb plot.
}
\description{
Each point shows the median dependency score for the queried genes within a given cell sub-lineage. The intervals show the 5 percent quantiles centered on the median, the interquartile ranges, and the 95 percent quantiles.
}
\examples{
make_sublineage(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_sublineage(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), highlight = TRUE)
make_sublineage(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_sublineage(input = list(type = 'gene', content = 'ROCK1'))
}
}
