% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_cellosaurus}
\alias{make_summary_cellosaurus}
\title{Make Cellosaurus Summary}
\usage{
make_summary_cellosaurus(
  data_cellosaurus = cellosaurus,
  input = list(),
  var = "ID"
)
}
\description{
Make Cellosaurus Summary
}
\examples{
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "SX") #sex
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "AG") #age
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "CC") \%>\% lit_linkr(data_gene_summary = gene_summary)
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "ATCC_url") #url
}
