% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_compound}
\alias{make_summary_compound}
\title{Make Compound Summary}
\usage{
make_summary_compound(
<<<<<<< HEAD
  data_universal_prism_meta = universal_prism_meta,
=======
  data_prism_meta = prism_meta,
>>>>>>> main
  input = list(),
  var = "name"
)
}
\description{
Make Compound Summary
}
\examples{
make_summary_compound(input = list(content = "aspirin"), var = "name")
}
