% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_lineage}
\alias{make_summary_lineage}
\title{Make Lineage Summary}
\usage{
make_summary_lineage(input = list(), var = "cell_line")
}
\description{
Make Lineage Summary
}
\examples{
make_summary_lineage(input = list(content = "Cervix"))
}
