% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_list}
\alias{make_summary_list}
\title{Make List Summary}
\usage{
make_summary_list(
  data_gene_summary = gene_summary,
  data_gene_location = gene_location,
  data_expression_meta = expression_meta,
  data_cellosaurus = cellosaurus,
  summary_len = 40,
  input = list()
)
}
\description{
Make List Summary
}
\examples{
make_summary_list(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
make_summary_list(input = list(type = "cell", content = c("HEL", "HEPG2")))
}
