% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_moa}
\alias{make_summary_moa}
\title{Make MOA Summary}
\usage{
make_summary_moa(
  data_universal_prism_meta = universal_prism_meta,
  input = list(),
  var = "name"
)
}
\description{
Make MOA Summary
}
\examples{
make_summary_moa(input = list(content = "cyclooxygenase inhibitor", content = "cyclooxygenase inhibitor"), var = "name")
make_summary_moa(input = list(content = "cyclooxygenase inhibitor", content = "cyclooxygenase inhibitor"), var = "moa")
}
