% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_protein}
\alias{make_summary_protein}
\title{Make Protein Summary}
\usage{
make_summary_protein(input = list(), summary_len = 40, ...)
}
\arguments{
\item{input}{A list containing a content variable.}

\item{var}{Variable that determines which text is returned}
}
\description{
The make_summary_protein function takes a gene as an input and returns summary text about its protein.
}
\examples{
make_summary_protein(input = list(content = "ROCK1"), var = "gene_name")
make_summary_protein(input = list(content = "ROCK2"), var = "protein_name")
make_summary_protein(input = list(content = "ROCK1"), var = "sequence")
}
