% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_text}
\alias{make_summary_text}
\title{Make Summary Text}
\usage{
make_summary_text(input = list(), summary_len = 40, ...)
}
\arguments{
\item{input}{A list containing a content variable.}

\item{var}{Variable that determines which text is returned}
}
\description{
This function takes one or many genes as inputs and returns summary list key summary information. This is a solution for make_summary_gene to make summaries of more than one gene.
}
\examples{
make_summary_text(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
make_summary_text(input = list(type = "cell", content = c("HEL", "HEPG2")))
make_summary_text(input = list(type = "gene", subtype = "pathway", query = 5887))
}
