% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_tissue}
\alias{make_tissue}
\title{Tissue Expression Plot}
\usage{
make_tissue(input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a tissue plot. If an error is thrown, then will return a bomb plot.
}
\description{
Human body tissue expression levels for the queried gene/s. The x-axis shows the normalized expression values and the y-axis shows the human body tissue.
}
\examples{
make_tissue(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_tissue(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_tissue(input = list(type = 'gene', content = 'ROCK1'))
}
}
