% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_top_table}
\alias{make_top_table}
\title{Top Table}
\usage{
make_top_table(
  toptable_data = master_top_table,
  toptable_data_cell = master_top_table_cell_line,
  upper = achilles_upper,
  gls = FALSE,
  input = list()
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a top Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_top_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a top Table
}
\examples{
make_top_table(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
\dontrun{
make_top_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
