% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_size_finder}
\alias{plot_size_finder}
\title{Function to obtain plot sizes}
\usage{
plot_size_finder(function_name)
}
\arguments{
\item{function_name}{Function name.}
}
\value{
A list with plot size
}
\description{
Function to obtain plot sizes
}
\examples{
plot_size_finder("make_ideogram")
}
