% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_ddh_d}
\alias{scale_color_ddh_d}
\title{Color scale constructor for discrete DDH colors}
\usage{
scale_color_ddh_d(
  palette = "gene",
  reverse = FALSE,
  shuffle = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in ddh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{shuffle}{Boolean indicating whether the colors should be shuffled}

\item{seed}{Boolean setting a seed to ensure consistent randomization}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE
or FALSE}
}
\description{
Color scale constructor for discrete DDH colors
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, cty, color = manufacturer)) +
  geom_point(size = 4) + scale_color_ddh_d()
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) + scale_color_ddh_d("protein", shuffle = TRUE, seed = 123L)

}
