% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_ddh_c}
\alias{scale_fill_ddh_c}
\title{Fill scale constructor for continuous DDH color palettes}
\usage{
scale_fill_ddh_c(palette = "gene", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ddh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE
or FALSE}
}
\description{
Fill scale constructor for continuous DDH color palettes
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, fill = Sepal.Width)) +
  geom_point(size = 4, shape = 21) + scale_fill_ddh_c()
ggplot(iris, aes(Sepal.Width, Sepal.Length, fill = Sepal.Width)) +
  geom_point(size = 4, shape = 21) +
  scale_fill_ddh_c("Gene_Protein", reverse = TRUE) + theme_ddh()

}
