% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_ddh_d}
\alias{scale_fill_ddh_d}
\title{Fill scale constructor for discrete DDH color palettes}
\usage{
scale_fill_ddh_d(
  palette = "gene",
  reverse = FALSE,
  shuffle = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in ddh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{shuffle}{Boolean indicating whether the colors should be shuffled}

\item{seed}{Boolean setting a seed to ensure consistent randomization}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE
or FALSE}
}
\description{
Fill scale constructor for discrete DDH color palettes
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, fill = class)) + geom_bar() + scale_fill_ddh_d()
ggplot(mpg, aes(class, fill = class)) + geom_bar() +
  scale_fill_ddh_d("protein", reverse = TRUE)
ggplot(mpg, aes(class, fill = class)) + geom_bar() +
  scale_fill_ddh_d("cell", shuffle = TRUE, seed = 1L)

}
