% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{send_report_message}
\alias{send_report_message}
\title{Send Report Message}
\usage{
send_report_message(
  first_name,
  last_name,
  email_address,
  input = list(),
  private
)
}
\arguments{
\item{first_name}{First name of the report addressee, which is used in the report email body}

\item{last_name}{Last name of the report addressee.}

\item{email_address}{Email address for the report to be delivered to}

\item{input}{A list containing type, query, and content variables.}

\item{private}{A Boolean determining which information to include in the report}
}
\description{
\code{send_report_message} generates a message that gets sent to the AWS simple queue service (SQS) for report generation
}
\details{
This is an emailing function that sends a message to the SQS queue for email generation
}
\examples{
send_report_message(first_name = "Matthew", last_name = "Hirschey", email_address = "hey_at_datadriventhypothesis.com", input = list(type = "gene", query = "ROCK1", content = "ROCK1"), private = TRUE)
\dontrun{
send_report_message(first_name = "Matthew", last_name = "Hirschey", email_address = "hey_at_datadriventhypothesis.com", input = list(type = "gene", query = "ROCK1", content = "ROCK1"), private = TRUE)
}
}
