% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_cell}
\alias{summary_cell}
\title{Cell summary}
\usage{
summary_cell(
  summary_table = expression_names,
  input = list(),
  var = "cell_line"
)
}
\description{
Cell summary
}
\examples{
summary_cell(input = list(content = "HEPG2"), var = "cell_line")
summary_cell(input = list(content = "HEPG2"), var = "lineage_subtype")
}
