% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_cellosaurus}
\alias{summary_cellosaurus}
\title{Cellosaurus}
\usage{
summary_cellosaurus(
  summary_table = cellosaurus,
  key_table = cellosaurus_key,
  input = list(),
  var = "ID"
)
}
\description{
Cellosaurus
}
\examples{
summary_cellosaurus(input = list(content = "HEPG2"), var = "SX") #sex
summary_cellosaurus(input = list(content = "HEPG2"), var = "AG") #age
summary_cellosaurus(input = list(content = "HEPG2"), var = "CC") \%>\% lit_linkr(summary_table = gene_summary)
summary_cellosaurus(input = list(content = "HEPG2"), var = "ATCC_url") #url
}
