% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_compound}
\alias{summary_compound}
\title{Compound Summary}
\usage{
summary_compound(summary_table = prism_meta, input = list(), var = "name")
}
\description{
Compound Summary
}
\examples{
summary_compound(input = list(query = "aspirin", content = "aspirin"), var = "name")
}
