% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_compound_list}
\alias{summary_compound_list}
\title{Compound list Summary}
\usage{
summary_compound_list(summary_table = prism_meta, input = list())
}
\description{
Compound list Summary
}
\examples{
summary_compound_list(input = list(content = c("aspirin", "carprofen")))
}
