% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_gene}
\alias{summary_gene}
\title{Gene Summary}
\usage{
summary_gene(
  summary_table = gene_summary,
  input = list(),
  var = "approved_symbol"
)
}
\description{
Gene Summary
}
\examples{
summary_gene(input = list(content = "ROCK1"), var = "approved_symbol")
summary_gene(input = list(content = "ROCK1"), var = "aka")
summary_gene(summary_table = gene_location, input = list(content = "ROCK1"), var = "cds_length")
}
