% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_lineage}
\alias{summary_lineage}
\title{Cell lineage summary}
\usage{
summary_lineage(
  summary_table = expression_names,
  input = list(),
  var = "cell_line"
)
}
\description{
Cell lineage summary
}
\examples{
summary_lineage(input = list(query = "Cervix"))
}
