% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_list}
\alias{summary_list}
\title{Summary List}
\usage{
summary_list(
  summary_gene = gene_summary,
  location_gene = gene_location,
  summary_cell = expression_meta,
  cell_meta = cellosaurus,
  summary_len = 40,
  input = list()
)
}
\description{
Summary List
}
\examples{
summary_list(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
summary_list(input = list(type = "cell", content = c("HEL", "HEPG2")))
}
