% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_moa}
\alias{summary_moa}
\title{MOA Summary}
\usage{
summary_moa(summary_table = prism_meta, input = list(), var = "name")
}
\description{
MOA Summary
}
\examples{
summary_moa(input = list(query = "cyclooxygenase inhibitor", content = "cyclooxygenase inhibitor"), var = "name")
summary_moa(input = list(query = "cyclooxygenase inhibitor", content = "cyclooxygenase inhibitor"), var = "moa")
}
