% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_pathway}
\alias{summary_pathway}
\title{Pathway Summary}
\usage{
summary_pathway(summary_table = pathways, input = list(), var = "pathway")
}
\description{
Pathway Summary
}
\examples{
summary_pathway(input = list(query = "1902965"), var = "data")
}
