% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summary_protein}
\alias{summary_protein}
\title{Protein summary}
\usage{
summary_protein(summary_table = proteins, input = list(), var = "gene_name")
}
\description{
Protein summary
}
\examples{
summary_protein(input = list(query = "ROCK1"), var = "gene_name")
summary_protein(input = list(query = "ROCK2"), var = "protein_name")
summary_protein(summary_table = proteins, input = list(query = "ROCK1"), var = "sequence")
}
