% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ddh}
\alias{theme_ddh}
\title{A ggplot theme which allow custom yet consistent styling of plots in the
Data-driven hypothesis web app.}
\usage{
theme_ddh(
  style = "default",
  base_size = 15,
  grid = "xy",
  axisline = "xy",
  axistitle = "xy",
  axistext = "xy",
  axis_num = "none",
  legend_num = FALSE,
  panelborder = FALSE,
  panelbackground = FALSE,
  margin = 2,
  ...
)
}
\arguments{
\item{style}{(string) Overall color style of text labels.
Either "default" or "light".}

\item{base_size}{(integer) Base point size}

\item{grid}{(string) Grid lines. Options include  "none" or
any combination of "X", "Y", "x" and "y".}

\item{axistitle}{(string) Axis titles. Options include "none" or
any combination of "X", "Y", "x" and "y".}

\item{axistext}{(string) Axis text labels for values or groups.
Options include "none" or any combination of "X", "Y", "x" and "y".}

\item{axis_num}{(string) Should axis text be formatted as monospaced? Set
to  x and y, respectively, in case numeric values are displayed. Options
include "none" or any combination of "X", "Y", "x" and "y".}

\item{legend_num}{(logical) Should legend text be formatted as monospaced?
Defaults to FALSE (no monospace font). Set to TRUE in case of numeric values.}

\item{panelborder}{(logical) Should a panel border be drawn?
Defaults to FALSE (no border). If TRUE it also adds tick marks to both axes.}

\item{margin}{(numeric) Should a margin of x be added to the plot?
Defaults to 0 (no margin by default).}

\item{...}{Other arguments passed to ggplot methods.}
}
\description{
A ggplot theme which allow custom yet consistent styling of plots in the
Data-driven hypothesis web app.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(diamonds, aes(cut)) + geom_bar() + theme_ddh()
ggplot(diamonds, aes(cut)) + geom_bar() +
  theme_ddh(style = "light", axisline = "none")
ggplot(diamonds, aes(cut)) + geom_bar() +
  theme_ddh(grid = "y", axisline = "none", panelborder = TRUE,
            panelbackground = TRUE, axis_num = "y", margin = 20)
}

}
