% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{title_extract}
\alias{title_extract}
\title{Extract Title Function}
\usage{
title_extract(fun, ...)
}
\arguments{
\item{fun}{Function name.}
}
\description{
Extract function title from documentation.
}
\examples{
title_extract("make_radial", package = ddh)
title_extract("make_umap_plot", package = ddh)
}
