% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{censor}
\alias{censor}
\title{Censor}
\usage{
censor(
  data_gene_master_top_table = gene_master_top_table,
  data_gene_censor = gene_censor,
  choice = FALSE,
  greater_than
)
}
\description{
Censor is used to remove genes from similarity table that are garbage (too many associations)
}
