% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{censor}
\alias{censor}
\title{Censor}
\usage{
censor(input = list(), choice = FALSE, greater_than = 300)
}
\description{
Censor is used to remove genes from similarity table that are garbage (too many associations). It does this by filtering out genes above (greater_than) a provided threshold.
}
\examples{
censor(input = list(type = 'gene', content = 'ROCK1'))
censor(input = list(type = 'gene', content = 'ROCK1'), choice = TRUE)
censor(input = list(type = 'gene', content = 'ROCK1'), choice = TRUE, greater_than = 100)
\dontrun{
censor(input = list(type = 'gene', content = 'ROCK1'))
}
}
