% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{download_ddh_data}
\alias{download_ddh_data}
\title{Function to download all DDH .Rds files}
\usage{
download_ddh_data(
  app_data_dir,
  file_name = NULL,
  overwrite = FALSE,
  privateMode = TRUE
)
}
\arguments{
\item{app_data_dir}{Data directory path.}

\item{file_name}{Optional file name to get a single file; default gets all files in bucket}

\item{overwrite}{Boolean that deletes the app_data_dir thereby removing files, before remaking dir and downloading all files}

\item{privateMode}{Boolean indicating if private data is required.}
}
\description{
Function to download all DDH .Rds files
}
