% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_ddh_data}
\alias{load_ddh_data}
\title{Function to Load All DDH Data Including .RDS Files and Colors}
\usage{
load_ddh_data(app_data_dir = NULL, release = "22Q2", privateMode = TRUE)
}
\arguments{
\item{privateMode}{}
}
\description{
Function to Load All DDH Data Including .RDS Files and Colors
}
