% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_ddh_data}
\alias{load_ddh_data}
\title{Function to Load All DDH Data Including .RDS Files and Colors}
\usage{
load_ddh_data(app_data_dir, file_name = NULL, privateMode = TRUE)
}
\arguments{
\item{app_data_dir}{Data directory path.}

\item{file_name}{Optional file name to get a single file; default loads all files}

\item{privateMode}{Boolean indicating if private data is required.}
}
\description{
Function to Load All DDH Data Including .RDS Files and Colors
}
