% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_ddh_data}
\alias{load_ddh_data}
\title{Function to Load All DDH Data Including .RDS Files and Colors}
\usage{
load_ddh_data(app_data_dir = NULL, release = "22Q2", privateMode = TRUE)
}
\arguments{
\item{app_data_dir}{Data directory path.}

\item{release}{Character vector indicating the release version.}

\item{privateMode}{Boolean indicating if private data is required.}
}
\description{
Function to Load All DDH Data Including .RDS Files and Colors
}
