% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_ddh_rds}
\alias{load_ddh_rds}
\title{Function to load all DDH .RDS files}
\usage{
load_ddh_rds(app_data_dir, object_name = NULL)
}
\arguments{
\item{app_data_dir}{Data directory path.}

\item{object_name}{Optional object name to load a single file; default loads all files}
}
\description{
Function to load all DDH .RDS files
}
