% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_ddh_rds}
\alias{load_ddh_rds}
\title{Function to load all DDH .RDS files}
\usage{
load_ddh_rds(app_data_dir = NULL, release = "22Q2", privateMode = TRUE)
}
\arguments{
\item{app_data_dir}{Data directory path.}

\item{release}{Character vector indicating the release version.}

\item{privateMode}{Boolean indicating if private data is required.}
}
\description{
Function to load all DDH .RDS files
}
