% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load_pdb}
\alias{load_pdb}
\title{Load PDB file}
\usage{
load_pdb(data_dir = app_data_dir, input = list())
}
\arguments{
\item{data_dir}{}

\item{input}{Expecting a list containing type and content variable.}
}
\value{
A PDB file.
}
\description{
Load PDB file
}
\examples{
load_pdb(input = list(content = c("ROCK1")))
load_pdb(input = list(content = c("ROCK3")))
load_pdb(input = list(content = c("ROCK1", "ROCK2")))
}
