% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_barcode}
\alias{make_barcode}
\title{Barcode Plot}
\usage{
make_barcode(input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing a content variable.}

\item{card}{A boolean that sets which image should be pulled for display}
}
\value{
If no error, then returns a barcode image url.
}
\description{
The barcode image shows the location of the query gene(s) on human chromosomes in the style of a barcode.
}
\examples{
make_barcode(input = list(type = "gene", query = "ROCK1", content = "ROCK1"))
make_barcode(input = list(content = "ROCK1"))
\dontrun{
make_barcode(input = list(type = "gene", content = "ROCK1"))
}
}
