% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_bipartite_graph}
\alias{make_bipartite_graph}
\title{Bipartite Graph Graph}
\usage{
make_bipartite_graph(
  toptable_data = master_top_table,
  bottomtable_data = master_bottom_table,
  protein_data = hmdb_proteins,
  metabolite_data = hmdb_metabolites,
  input = list(),
  censor = character(),
  collapsed = TRUE,
  threshold = 10,
  corrType = "Positive"
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a bipartite graph graph. If an error is thrown, then will return an empty graph.
}
\description{
\code{make_bipartite_graph} returns an image of ...
}
\details{
This is a graph function that takes a gene name and returns a bipartite graph graph
}
\examples{
make_bipartite_graph(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_bipartite_graph(input = list(type = "gene", content = "ROCK1"), collapsed = TRUE, threshold = 10, corrType = "Positive")
make_bipartite_graph(input = list(type = "gene", content = c("ROCK1", "ROCK2")), collapsed = TRUE, threshold = 10, corrType = "Positive", censor = c("ADP", "Adenosine triphosphate"))
make_bipartite_graph(input = list(type = "compound", content = "adp"))
\dontrun{
make_bipartite_graph(input = list(type = 'gene', content = 'ROCK1'))
}
}
