% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cca_genes}
\alias{make_cca_genes}
\title{Co-essentiality Pathway Plot}
\usage{
make_cca_genes(
  input = list(),
  n_features = 5,
  gset = NULL,
  facet_by_geneset = FALSE,
  ...
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a barplot. If an error is thrown, then will return a bomb plot.
}
\description{
Top pathway associated for the queried gene/s. The x-axis shows the first canonical correlation between the gene and the pathway (range 0-1).
}
\examples{
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), gset = "GOBP")
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), gset = c("GOBP", "C2"))
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
make_cca_genes(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')), n_features = 20)
}
