% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cell_drugs_table}
\alias{make_cell_drugs_table}
\title{Cell Drugs Table}
\usage{
make_cell_drugs_table(
  table_data = prism_long,
  cell_meta = expression_meta,
  input = list()
)
}
\description{
Cell Drugs Table
}
