% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_cell_sim_table}
\alias{make_cell_sim_table}
\title{Cell Similarity Table}
\usage{
make_cell_sim_table(
  input = list(),
  similarity = "dependency",
  bonferroni_cutoff = 0.05
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a cell sim Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_cell_sim_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a cell sim Table
}
\examples{
make_cell_sim_table(input = list(type = "cell", content = "HEL"))
make_cell_sim_table(input = list(type = "cell", content = c("HEL", "LS513")))
\dontrun{
make_cell_sim_table(input = list(type = "cell", content = "HEL"))
}
}
