% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellanatogram}
\alias{make_cellanatogram}
\title{Cellanatogram Plot}
\usage{
make_cellanatogram(input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cellanatogram plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_cellanatogram} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a cellanatogram plot
}
\examples{
make_cellanatogram(input = list(type = "gene", content = c("ROCK2")))
make_cellanatogram(input = list(type = "gene", content = c("ROCK2")), card = TRUE)
make_cellanatogram(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
\dontrun{
make_cellanatogram(input = list(type = 'gene', content = 'ROCK2'))
}
}
