% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellanatogramfacet}
\alias{make_cellanatogramfacet}
\title{Cell Anatogram Facet}
\usage{
<<<<<<< HEAD
make_cellanatogramfacet(data_gene_subcell = gene_subcell, input = list())
=======
make_cellanatogramfacet(data_subcell = subcell, input = list())
>>>>>>> main
}
\description{
Cell Anatogram Facet
}
\examples{
make_cellanatogramfacet(input = list(type = "gene", content = c("BRCA1", "ROCK2")))
}
