% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellanatogramfacet}
\alias{make_cellanatogramfacet}
\title{Cell Anatogram Facet}
\usage{
make_cellanatogramfacet(input = list())
}
\description{
Cell Anatogram Facet
}
\examples{
make_cellanatogramfacet(input = list(type = "gene", content = c("BRCA1", "ROCK2")))
}
