% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_celldeps}
\alias{make_celldeps}
\title{Cell Dependencies Plot}
\usage{
make_celldeps(
  celldeps_data = achilles_long,
  prism_data = prism_long,
  expression_data = expression_names,
  input = list(),
  mean,
  card = FALSE,
  lineplot = FALSE,
  scale = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a celldeps plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_celldeps} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a celldeps plot
}
\examples{
make_celldeps(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_celldeps(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_celldeps(input = list(type = 'gene', content = 'ROCK1'))
}
}
