% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellexpression}
\alias{make_cellexpression}
\title{Cell Expression Plot}
\usage{
make_cellexpression(
<<<<<<< HEAD
  data_universal_expression_long = universal_expression_long,
  data_cell_expression_names = cell_expression_names,
  data_universal_stats_summary = universal_stats_summary,
=======
  data_expression_long = expression_long,
  data_expression_names = expression_names,
  data_gene_expression_upper = gene_expression_upper,
  data_gene_expression_lower = gene_expression_lower,
  data_protein_expression_upper = protein_expression_upper,
  data_protein_expression_lower = protein_expression_lower,
  data_mean_virtual_gene_expression = mean_virtual_gene_expression,
  data_mean_virtual_protein_expression = mean_virtual_protein_expression,
>>>>>>> main
  input = list(),
  var = "gene",
  card = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cellexpression plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_cellexpression} returns an image of ...
}
\details{
Each point shows the ranked expression value across CCLE cell lines.
}
\examples{
make_cellexpression(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cellexpression(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cellexpression(input = list(type = 'gene', content = 'ROCK1'))
}
}
