% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellgeneprotein}
\alias{make_cellgeneprotein}
\title{Cellgeneprotein Plot}
\usage{
make_cellgeneprotein(
  expression_data = expression_long,
  expression_join = expression_names,
  input = list(),
  card = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cellgeneprotein plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_cellgeneprotein} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a plot that compares this makes gene v. protein expression levels in a cell
}
\examples{
make_cellgeneprotein(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cellgeneprotein(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cellgeneprotein(input = list(type = 'gene', content = 'ROCK1'))
}
}
