% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cellgeneprotein}
\alias{make_cellgeneprotein}
\title{Gene Expression versus Protein Expression}
\usage{
make_cellgeneprotein(input = list(), card = FALSE)
}
\description{
Each point shows the gene expression value compared to the protein expression value for gene within a given cell line. The Pearson correlation coefficient and the p-values are provided in the top-left corner of the plot.
}
\examples{
make_cellgeneprotein(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cellgeneprotein(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cellgeneprotein(input = list(type = 'gene', content = 'ROCK1'))
}
}
