% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cluster_enrich}
\alias{make_cluster_enrich}
\title{Cluster Enrich Plot}
\usage{
make_cluster_enrich(input = list(), ontology = "BP", num_terms = 20)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cluster enrich plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_cluster_enrich} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a cluster enrich plot
}
\examples{
make_cluster_enrich(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cluster_enrich(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_cluster_enrich(input = list(type = 'gene', content = 'ROCK1'))
}
}
