% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_cluster_enrichment}
\alias{make_cluster_enrichment}
\title{Amino Acid Signature Enrichment Plot}
\usage{
make_cluster_enrichment(input = list(), ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a cluster enrich plot. If an error is thrown, then will return a bomb plot.
}
\description{
Enriched pathways for the proteins in the selected cluster.
}
\examples{
make_cluster_enrichment(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_cluster_enrichment(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
\dontrun{
make_cluster_enrichment(input = list(type = 'gene', content = 'ROCK1'))
}
}
