% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_clustering_table}
\alias{make_clustering_table}
\title{Clustering Table}
\usage{
make_clustering_table(
<<<<<<< HEAD
  data_gene_signature_clusters = gene_signature_clusters,
  data_gene_signatures = gene_signatures,
  data_gene_signature_cluster_names = gene_signature_cluster_names,
=======
  data_sequence_clusters = sequence_clusters,
  data_signatures = signatures,
  data_protein_cluster_names = data_hmdb_proteins_protein_cluster_names,
>>>>>>> main
  input = list(),
  cluster = FALSE,
  show_signature = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a clustering Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_clustering_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a clustering Table
}
\examples{
make_clustering_table(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_clustering_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
