% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_correlation}
\alias{make_correlation}
\title{Co-essentiality Correlation Plot}
\usage{
make_correlation(
  data_achilles_cor_nest = achilles_cor_nest,
  data_prism_cor_nest = prism_cor_nest,
  data_achilles_upper = achilles_upper,
  data_achilles_lower = achilles_lower,
  data_prism_cor_upper = prism_cor_upper,
  data_prism_cor_lower = prism_cor_lower,
  data_mean_virtual_achilles = mean_virtual_achilles,
  data_mean_virtual_prism_cor = mean_virtual_prism_cor,
  input = list(),
  card = FALSE,
  scale = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a correlation plot. If an error is thrown, then will return a bomb plot.
}
\description{
Each point shows the ranked correlation value ordered from high to low for each query. Correlation values outside the solid gray lines indicate the gene has a correlation value greater than the mean.
}
\examples{
make_correlation(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_correlation(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_correlation(input = list(type = 'gene', content = 'ROCK1'))
}
}
