% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_dep_table}
\alias{make_dep_table}
\title{Dependency Table}
\usage{
make_dep_table(
  achilles_data = achilles_long,
  prism_data = prism_long,
  prism_summary = prism_names,
  expression_data = expression_names,
  input = list(),
  var = "gene"
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a dep Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_dep_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a dep Table
}
\examples{
make_dep_table(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_dep_table(input = list(type = 'compound', query = 'aspirin', content = 'aspirin'))
make_dep_table(input = list(type = 'cell', query = 'HEPG2', content = 'HEPG2'))
\dontrun{
make_dep_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
