% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_dep_table}
\alias{make_dep_table}
\title{Dependency Table}
\usage{
make_dep_table(
<<<<<<< HEAD
  data_universal_achilles_long = universal_achilles_long,
  data_universal_prism_long = universal_prism_long,
  data_compound_prism_names = compound_prism_names,
  data_cell_expression_names = cell_expression_names,
  data_universal_gene_summary = universal_gene_summary,
=======
  data_achilles_long = achilles_long,
  data_prism_long = prism_long,
  data_prism_names = prism_names,
  data_expression_names = expression_names,
  data_gene_summary = gene_summary,
>>>>>>> main
  input = list(),
  var = "gene"
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a dep Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_dep_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a dep Table
}
\examples{
make_dep_table(input = list(type = 'gene', content = 'ROCK1'))
make_dep_table(input = list(type = 'compound', content = 'aspirin'))
make_dep_table(input = list(type = 'cell', content = 'HEPG2'))
\dontrun{
make_dep_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
