% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_drug_genes_cor_table}
\alias{make_drug_genes_cor_table}
\title{Drug Genes Cor Table}
\usage{
make_drug_genes_cor_table(
<<<<<<< HEAD
  data_compound_genes_cor_table = compound_genes_cor_table,
=======
  data_drug_genes_cor_table = drug_genes_cor_table,
>>>>>>> main
  drug
)
}
\description{
\code{make_drug_genes_cor_table} takes a drug name and returns a table of genes
}
\details{
This is a table function that takes a drug name and returns a table of genes
}
