% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_drug_genes_cor_table}
\alias{make_drug_genes_cor_table}
\title{Drug Genes Cor Table}
\usage{
make_drug_genes_cor_table(table_data = drug_genes_cor_table, drug)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a drug genes cor Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_drug_genes_cor_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a drug genes cor Table
}
\examples{
make_drug_genes_table(input = list(content = "aspirin"))
\dontrun{
make_drug_genes_table(input = list(content = "aspirin"))
}
}
