% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_expdep}
\alias{make_expdep}
\title{Gene Dependency versus Expression}
\usage{
make_expdep(
  expression_data = expression_long,
  celldeps_data = achilles_long,
  expression_join = expression_names,
  plot_se = TRUE,
  input = list(),
  card = FALSE
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a expdep plot. If an error is thrown, then will return a bomb plot.
}
\description{
Each point shows the dependency value compared to the expression value for gene within a given cell line. Gray area indicates dependency values that are between -1 and 1.
}
\examples{
make_expdep(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_expdep(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_expdep(input = list(type = 'gene', content = 'ROCK1'))
}
}
