% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_female_anatogram}
\alias{make_female_anatogram}
\title{Female Anatogram Plot}
\usage{
make_female_anatogram(input = list(), anatogram = "female", card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a female anatogram plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_female_anatogram} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a female anatogram plot
}
\examples{
make_female_anatogram(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_female_anatogram(input = list(type = 'gene', content = c('ROCK1', 'ROCK2')))
make_female_anatogram(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_female_anatogram(input = list(type = 'gene', content = 'ROCK1'))
}
}
