% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_gene_drugs_table}
\alias{make_gene_drugs_table}
\title{Gene Drugs Table}
\usage{
make_gene_drugs_table(table_data = gene_drugs_table, input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a drug genes Table. If an error is thrown, then will return an empty table.
}
\description{
This is a table function that takes a gene name and returns a drug genes Table
}
\examples{
make_gene_drugs_table(inpuit = list(type = "gene", content = "ROCK1"))
\dontrun{
make_gene_drugs_table(inpuit = list(type = "gene", content = "ROCK1"))
}
}
