% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_graph}
\alias{make_graph}
\title{Graph Graph}
\usage{
make_graph(
  toptable_data = master_top_table,
  bottomtable_data = master_bottom_table,
  compound_data = prism_cor_nest,
  input = list(),
  threshold = 10,
  deg = 2,
  corrType = "Positive",
  displayHeight = "90vh",
  displayWidth = "100\%",
  tooltipLink = FALSE,
  card = FALSE
)
}
\arguments{
\item{toptable_data}{A tibble of genes and their associated top correlated genes}

\item{bottomtable_data}{A tibble of genes and their associated bottom correlated genes}

\item{input}{A list containing character vector of gene_symbols used to create network graph}

\item{threshold}{A numerical representing the number of genes to pull from top and bottom tables}

\item{deg}{A numerical representing the minimum number of connections for a gene to be connected to the network}

\item{corrType}{A string that describes what type of correlations to include, options are: "Positive and Negative", "Positive", or "Negative"}

\item{displayHeight}{Default to "90vh". The height of the network in pixels("500px"), as a percentage("100%"), or as a percentage of the viewport("70vh", where 70 represents 70% of the viewport)}

\item{displayWidth}{Default to "100%". The width of the network in pixels("500px"), as a percentage("100%"), or as a percentage of the viewport("70vh", where 70 represents 70% of the viewport)}

\item{tooltipLink}{Boolean to denote whether or not to include a link in the tooltip for a gene. Default to false.}
}
\value{
Outputs a complete network graph. If an error is thrown, then will return an empty graph.
}
\description{
\code{make_graph} returns an image of ...
}
\details{
This is a graph function that takes a gene name and returns a graph graph
}
\examples{
make_graph(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_graph(input = list(type = "gene", content = "ROCK1"))
make_graph(input = list(type = "gene", content = "ROCK1"), card = TRUE)
make_graph(input = list(type = "gene", content = "ROCK1"), corrType = "Positive")
make_graph(input = list(type = "pathway", query = "1902965", content = c("RDX", "ROCK2", "DTX3L", "MSN", "SORL1", "EZR")), corrType = "Positive")
make_graph(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
make_graph(input = list(type = "gene", content = "DTX3L"), corrType = "Negative") # disconnected query gene
make_graph(input = list(type = "compound", content = "aspirin"), corrType = "Negative")
\dontrun{
make_graph(input = list(type = 'gene', content = 'ROCK1'))
}
}
