% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_male_anatogram}
\alias{make_male_anatogram}
\title{Male Anatogram Plot}
\usage{
make_male_anatogram(anatogram = "male", input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{

}
\description{
This is a plot function that takes a gene name and returns a male anatogram plot
}
\examples{
make_male_anatogram(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_male_anatogram(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_male_anatogram(input = list(type = 'gene', content = 'ROCK1'))
}
}
