% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_metabolite_table}
\alias{make_metabolite_table}
\title{Metabolite Table}
\usage{
make_metabolite_table(input = list(), collapse = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a metabolite Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_metabolite_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a metabolite Table
}
\examples{
make_metabolite_table(input = list(type = 'gene', content = 'ROCK1'))
make_metabolite_table(input = list(type = 'gene', content = 'ROCK1'), collapse = TRUE)
make_metabolite_table(input = list(type = "gene", content = "ROCK2"))
make_metabolite_table(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
make_metabolite_table(input = list(type = "compound", content = "NAD"))
make_metabolite_table(input = list(type = "cell", content = "HEPG2"))
\dontrun{
make_metabolite_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
