% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_molecular_features}
\alias{make_molecular_features}
\title{Molecular Features Bar Plot}
\usage{
make_molecular_features(input = list(), n_features = 5, ...)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a barplot. If an error is thrown, then will return a bomb plot.
}
\description{
Top molecular features associated to the queried gene/s ablation. The x-axis shows the log2 fold change (reference group: sensitive).
}
\examples{
make_molecular_features(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_molecular_features(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')))
make_molecular_features(input = list(type = 'gene', query = 'ROCK1', content = c('ROCK1', 'ROCK2')), n_features = 20)
}
