% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_pathway_genes}
\alias{make_pathway_genes}
\title{Pathway Genes Table}
\usage{
make_pathway_genes(input = list())
}
\description{
\code{make_pathway_genes} returns a table of genes in a queried pathway
}
\details{
This is a table function that takes a gene_set id in the query slot and returns a sub-table of genes that your gene set contains
#'
}
\examples{
make_pathway_genes(input = list(type = 'gene', subtype = 'pathway', query = 5887))
}
