% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_pathway_genes}
\alias{make_pathway_genes}
\title{Pathway Genes Table}
\usage{
make_pathway_genes(table_name = pathways, table_join = gene_summary, go_id)
}
\description{
Pathway Genes Table
}
\examples{
make_pathway_genes(go_id = "1902965")
}
