% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_pathway_list}
\alias{make_pathway_list}
\title{Pathway List Table}
\usage{
make_pathway_list(table_name = pathways, input = list())
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a pathway list table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_pathway_list} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a pathway list table
}
\examples{
make_pathway_list(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
\dontrun{
make_pathway_list(input = list(type = 'gene', content = 'ROCK1'))
}
}
