% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_protein_domain_plot}
\alias{make_protein_domain_plot}
\title{Protein Domain Plot}
\usage{
make_protein_domain_plot(
  input = list(),
  domain_data = protein_domains,
  dom_var = NULL,
  ptm_var = NULL
)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a protein domain plot. If an error is thrown, then will return a bomb plot.
}
\description{
Rectangles represent the locations and size of named protein domains, while black shaped elements represent PTMs. Horizontal line(s) indicate the length of one or more selected proteins.
}
\examples{
make_protein_domain_plot(input = list(content = "ROCK2"), dom_var = "Protein kinase", ptm_var = "N-acetylserine")
make_protein_domain_plot(input = list(content = c("ROCK1", "ROCK2")), dom_var = "Protein kinase", ptm_var = "N-acetylserine")
\dontrun{
make_protein_domain_plot(input = list(content = "ROCK2"), dom_var = "Protein kinase", ptm_var = "N-acetylserine")
}
}
