% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_pubmed}
\alias{make_pubmed}
\title{Pubmed Plot}
\usage{
make_pubmed(pubmed_data = pubmed, input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a pubmed plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_pubmed} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a pubmed plot
}
\examples{
make_pubmed(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_pubmed(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
make_pubmed(input = list(type = 'compound', content = 'aspirin'))
\dontrun{
make_pubmed(input = list(type = 'gene', content = 'ROCK1'))
}
}
