% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_pubmed_table}
\alias{make_pubmed_table}
\title{Pubmed Table}
\usage{
<<<<<<< HEAD
make_pubmed_table(data_universal_pubmed = universal_pubmed, input = list())
=======
make_pubmed_table(data_pubmed = pubmed, input = list())
>>>>>>> main
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}
}
\value{
If no error, then returns a pubmed Table. If an error is thrown, then will return an empty table.
}
\description{
\code{make_pubmed_table} returns an image of ...
}
\details{
This is a table function that takes a gene name and returns a pubmed Table
}
\examples{
make_pubmed_table(input = list(type = 'gene', content = 'ROCK1'))
\dontrun{
make_pubmed_table(input = list(type = 'gene', content = 'ROCK1'))
}
}
