% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_structure}
\alias{make_structure}
\title{Protein Structure Plot}
\usage{
make_structure(input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a structure plot. If an error is thrown, then will return a bomb plot.
}
\description{
Alpha Fold predicted structure rendered in a ribbon diagram.
}
\examples{
make_structure(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_structure(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_structure(input = list(type = 'gene', content = 'ROCK1'))
}
}
