% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_structure3d_table}
\alias{make_structure3d_table}
\title{Plot for 3D protein structure table}
\usage{
make_structure3d_table(
  pdb_ids = uniprot_pdb_table,
  protein_data = proteins,
  input = list()
)
}
\description{
@importFrom magrittr %>%
}
