% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_cell}
\alias{make_summary_cell}
\title{Make Cell Summary}
\usage{
make_summary_cell(
  data_cell_expression_names = cell_expression_names,
  input = list(),
  var = "cell_line"
)
}
\description{
Make Cell Summary
}
\examples{
make_summary_cell(input = list(content = "HEPG2"), var = "cell_line")
make_summary_cell(input = list(content = "HEPG2"), var = "lineage_subtype")
}
