% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_cellosaurus}
\alias{make_summary_cellosaurus}
\title{Make cell_osaurus Summary}
\usage{
make_summary_cellosaurus(
  data_cell_osaurus = cell_osaurus,
  input = list(),
  var = "ID"
)
}
\description{
Make cell_osaurus Summary
}
\examples{
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "SX") #sex
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "AG") #age
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "CC") \%>\% lit_linkr(data_universal_gene_summary = universal_gene_summary)
make_summary_cellosaurus(input = list(content = "HEPG2"), var = "ATCC_url") #url
}
