% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_compound_list}
\alias{make_summary_compound_list}
\title{Make Compound List Summary}
\usage{
make_summary_compound_list(
  data_universal_prism_meta = universal_prism_meta,
  input = list()
)
}
\description{
Make Compound List Summary
}
\examples{
make_summary_compound_list(input = list(content = c("aspirin", "carprofen")))
}
