% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_gene}
\alias{make_summary_gene}
\title{Make Gene Summary}
\usage{
make_summary_gene(
<<<<<<< HEAD
  data_universal_gene_summary = universal_gene_summary,
=======
  data_gene_summary = gene_summary,
>>>>>>> main
  input = list(),
  var = "approved_symbol"
)
}
\description{
Make Gene Summary
}
\examples{
make_summary_gene(input = list(content = "ROCK1"), var = "approved_symbol")
make_summary_gene(input = list(content = "ROCK1"), var = "aka")
<<<<<<< HEAD
make_summary_gene(input = list(content = "ROCK1"), var = "cds_length")
=======
make_summary_gene(data_gene_summary = gene_location, input = list(content = "ROCK1"), var = "cds_length")
>>>>>>> main
}
