% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_gene}
\alias{make_summary_gene}
\title{Make Gene Summary}
\usage{
make_summary_gene(input = list(), var = "approved_symbol")
}
\arguments{
\item{input}{A list containing a content variable.}

\item{var}{Variable that determines which text is returned}
}
\description{
The make_gene_summary function takes a gene as an input and returns summary text.
}
\examples{
make_summary_gene(input = list(content = "ROCK1"), var = "approved_symbol")
make_summary_gene(input = list(content = "ROCK1"), var = "approved_name")
make_summary_gene(input = list(content = "ROCK1"), var = "entrez_summary")
make_summary_gene(input = list(content = c("ROCK1", "ROCK2")), var = "approved_name")
}
