% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_list}
\alias{make_summary_list}
\title{Make List Summary}
\usage{
make_summary_list(
  data_universal_gene_summary = universal_gene_summary,
  data_gene_location = gene_location,
  data_cell_expression_meta = cell_expression_meta,
  data_cell_osaurus = cell_osaurus,
  summary_len = 40,
  input = list()
)
}
\arguments{
\item{input}{A list containing a content variable.}

\item{var}{Variable that determines which text is returned}
}
\description{
The make_summary_list function takes one or many genes as inputs and returns summary list key summary information. This is a solution for make_summary_gene to make summaries of more than one gene.
}
\examples{
make_summary_list(input = list(type = "gene", content = c("ROCK1", "ROCK2")))
make_summary_list(input = list(type = "cell", content = c("HEL", "HEPG2")))
}
