% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_metabolite}
\alias{make_summary_metabolite}
\title{Make Metabolite Summary}
\usage{
make_summary_metabolite(
  data_compound_hmdb_names = compound_hmdb_names,
  input = list(),
  var = "name"
)
}
\description{
Make Metabolite Summary
}
\examples{
make_summary_metabolite(input = list(content = "citrate"), var = "name")
}
