% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_pathway}
\alias{make_summary_pathway}
\title{Make Pathway Summary}
\usage{
make_summary_pathway(data_pathways = pathways, input = list(), var = "pathway")
}
\description{
Make Pathway Summary
}
\examples{
make_summary_pathway(input = list(content = "1902965"), var = "data")
}
