% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_pathway}
\alias{make_summary_pathway}
\title{Make Pathway Summary}
\usage{
make_summary_pathway(input = list(), var = "gs_description")
}
\arguments{
\item{input}{A list containing a content variable.}

\item{var}{Variable that determines which text is returned}
}
\description{
The make_pathway_summary function takes an id as a character input and returns summary information of the pathway.
}
\examples{
make_summary_pathway(input = list(query = 5887), var = "gs_name")
make_summary_pathway(input = list(query = 5887), var = "gs_cat")
make_summary_pathway(input = list(query = 5887), var = "gs_description")
make_summary_pathway(input = list(query = 5887), var = "pathway_size")
}
