% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_summary_protein}
\alias{make_summary_protein}
\title{Make Protein Summary}
\usage{
make_summary_protein(
  data_proteins = proteins,
  input = list(),
  var = "gene_name"
)
}
\description{
Make Protein Summary
}
\examples{
make_summary_protein(input = list(content = "ROCK1"), var = "gene_name")
make_summary_protein(input = list(content = "ROCK2"), var = "protein_name")
make_summary_protein(data_proteins = proteins, input = list(content = "ROCK1"), var = "sequence")
}
