% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_tissue}
\alias{make_tissue}
\title{Tissue Plot}
\usage{
make_tissue(tissue_data = tissue, input = list(), card = FALSE)
}
\arguments{
\item{input}{Expecting a list containing type and content variable.}

\item{card}{A boolean that sets whether the plot should be scaled down to be a card}
}
\value{
If no error, then returns a tissue plot. If an error is thrown, then will return a bomb plot.
}
\description{
\code{make_tissue} returns an image of ...
}
\details{
This is a plot function that takes a gene name and returns a tissue plot
}
\examples{
make_tissue(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'))
make_tissue(input = list(type = 'gene', query = 'ROCK1', content = 'ROCK1'), card = TRUE)
\dontrun{
make_tissue(input = list(type = 'gene', content = 'ROCK1'))
}
}
