% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{setup_graph}
\alias{setup_graph}
\title{Setup graph parameters}
\usage{
setup_graph(
  toptable_data = master_top_table,
  bottomtable_data = master_bottom_table,
  compound_data = prism_cor_nest,
  input_list = list(),
  setup_threshold,
  setup_corrType,
  cell_line_similarity = "dependency",
  cell_dep = cell_line_dep_sim,
  cell_exp = cell_line_exp_sim,
  bonferroni_cutoff = 0.05
)
}
\description{
Setup graph parameters
}
