% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{setup_graph}
\alias{setup_graph}
\title{Setup graph parameters}
\usage{
setup_graph(
  data_master_top_table = master_top_table,
  data_master_bottom_table = master_bottom_table,
  data_prism_cor_nest = prism_cor_nest,
  input_list = list(),
  setup_threshold,
  setup_corrType,
  cell_line_similarity = "dependency",
  data_cell_line_dep_sim = cell_line_dep_sim,
  data_cell_line_exp_sim = cell_line_exp_sim,
  bonferroni_cutoff = 0.05
)
}
\description{
Setup graph parameters
}
